+++ src/main/java/conflux/web3j/Web3j.java
@@ -213,6 +213,12 @@ class Web3j implements Web3 {
 	}
 
 	@Override
+	public Request<BigInteger, BigIntResponse> txpoolNextNonce(Address address) {
+		return new Request<>(this.service, "txpool_nextNonce", BigIntResponse.class, address)
+				.withRetry(this.retry, this.intervalMillis);
+	}
+
+	@Override
 	public Request<String, StringResponse> sendRawTransaction(String hexEncoded) {
 		return new Request<>(this.service, "cfx_sendRawTransaction", StringResponse.class, hexEncoded)
 				.withRetry(this.retry, this.intervalMillis);
@@ -397,13 +403,25 @@ class Web3j implements Web3 {
 	}
 
 	@Override
+	public Request<List<String>, StringListResponse.Response> rpcModules() {
+		return new Request<>(this.service, "cfx_openedMethodGroups", StringListResponse.Response.class);
+	}
+
+	@Override
+	public Request<PoSEconomics, PoSEconomics.Response> getPoSEconomics() {
+		return new Request<>(this.service, "cfx_getPoSEconomics", PoSEconomics.Response.class);
+	}
+
+	@Override
+	public Request<PoSEpochRewards, PoSEpochRewards.Response> getPoSRewardByEpoch(Address address, Epoch epoch) {
+		return new Request<>(this.service, "cfx_getPoSRewardByEpoch", PoSEpochRewards.Response.class);
+	}
+
+	@Override
 	public <T,R extends Response<?> & HasValue<T>> Request<T, R> getCustomizedRequest(Class<R> responseType, String method, Object... params){
 		return new Request<>(this.service, method, responseType, params)
 				.withRetry(this.retry, this.intervalMillis);
 	}
-
-
-
 
 	@Override
 	public Flowable<NewHeadsNotification> subscribeNewHeads() {