+++ src/main/java/conflux/web3j/response/StringListResponse.java
@@ -1,0 +1,5 @@
+package conflux.web3j.response;
+
+public class StringListResponse {
+    public static class Response extends CfxListResponse<String> {}
+}