+++ src/main/java/conflux/web3j/response/UsedGasAndCollateral.java
@@ -9,6 +9,7 @@ public class UsedGasAndCollateral {
 	public static class Response extends CfxResponse<UsedGasAndCollateral> {}
 	
 	private String gasUsed;
+	private String gasLimit;
 	private String storageCollateralized;
 	
 	public BigInteger getGasUsed() {
@@ -17,6 +18,14 @@ public class UsedGasAndCollateral {
 	
 	public void setGasUsed(String gasUsed) {
 		this.gasUsed = gasUsed;
+	}
+
+	public BigInteger getGasLimit() {
+		return Numeric.decodeQuantity(this.gasLimit);
+	}
+
+	public void setGasLimit(String gasLimit) {
+		this.gasLimit = gasLimit;
 	}
 	
 	public BigInteger getStorageCollateralized() {